# openapi.yml
openapi: "3.1.0"
info:
  title: Geode Index
  description: Documentation for the Geode Index API
  version: "1.0.0"
servers:
  - url: https://api.geode-sdk.org/
...
paths:
...
  /v1/mods:
    get:
      tags:
        - mods
      summary: Search all mods, paginated
      description: Returns a list of mods, paginated. Does not include about.md and changelog.md. Does not include dependencies and incompatibilities.

      parameters:
        - name: query
          in: query
          description: Search query
          schema:
            type: string
        - name: gd
          in: query
          description: Geometry Dash version
          required: true
          schema:
            $ref: "#/components/schemas/GDVersionString"
        - name: platforms
          in: query
          description: Platforms that mods have to support, comma separated [win,android32,android64,mac,ios]
          example: "win,android32,android64"
          schema:
            type: string
        - name: tags
          in: query
          description: Tags that mods have to have, comma separated. Can only use allowed tags.
          example: "gameplay,enhancement"
          schema:
            type: string
        - name: featured
          in: query
          description: Filter by featured mods
          schema:
            type: boolean
        - name: sort
          in: query
          description: Sort by a specific enum field, in descending order. Default is "downloads"
          example: "downloads"
          schema:
            type: enum
            enum:
              - downloads
              - recently_updated
              - recently_published
        - name: developer
          in: query
          description: Filter by developer username
          schema:
            type: string
        - name: status
          in: query
          description: Filter by mod version status (rejected is admin only)
          schema:
            $ref: "#/components/schemas/ModVersionStatus"
        - name: pending_validation
          in: query
          description: Filter by pending validation status (admin only)
          schema:
            type: boolean
        - $ref: "#/components/parameters/GeodeVersionQuery"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mod"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - mods
      summary: Create a new mod
      description: This endpoint is only used for adding a new mod. For creating a new version, use the POST /mods/{id}/versions endpoint.
      security:
        - bearerAuth: []

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                download_link:
                  type: string
                  description: The download URL for the .geode file. You can use a Github Release for this.
                  examples:
                    - "https://github.com/geode-sdk/NodeIDs/releases/download/v1.0.0/geode.node-ids.geode"
                    - "https://github.com/geode-sdk/DevTools/releases/download/v1.3.0/geode.devtools.geode"

      responses:
        "204":
          description: No Content (Mod created)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/mods/{id}:
    get:
      tags:
        - mods
      summary: Get info for a mod
      description: Returns info for the mod, without dependencies and incompatibilities.
      parameters:
        - $ref: "#/components/parameters/ModID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mod"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags:
        - mods
      summary: Update a mod (admin only)
      description: At the moment it's only used to set a mod to featured / not featured
      security:
        - bearerAuth: []

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                featured:
                  type: boolean
                  description: Whether the mod should be featured or not
                  example: true

      parameters:
        - $ref: "#/components/parameters/ModID"
      responses:
        "204":
          description: No Content (Mod updated)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/mods/{id}/logo:
    get:
      tags:
        - mods
      summary: Get the logo for a mod
      description: Returns the logo for the mod
      parameters:
        - $ref: "#/components/parameters/ModID"
      responses:
        "200":
          description: Returns the mod logo
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/mods/updates:
    get:
      tags:
        - mods
      summary: Get available mod updates
      description: Gets latest versions of the requested mods

      parameters:
        - name: platform
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Platform"
        - name: ids
          in: query
          required: true
          schema:
            type: string
            description: Semicolon separated list of mod IDs
            example: "geode.devtools;geode.nodeids"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: "null"
                  payload:
                    type: array
                    items: "#/components/schemas/ModUpdate"
...
  /v1/loader/versions:
    get:
      tags:
        - loader
      summary: Get all loader versions, paginated
      description: Returns a paginated list of all loader versions. This list is sorted by when the version was added, not by their version number.

      parameters:
        - name: gd
          in: query
          description: Geometry Dash version
          required: false
          schema:
            $ref: "#/components/schemas/GDVersionString"
        - name: platform
          in: query
          description: Platform to filter version by [win, android, mac, ios]
          required: false
          schema:
            $ref: "#/components/schemas/Platform"
        - name: prerelease
          in: query
          description: If prerelease builds should be included in results
          required: false
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoaderVersion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - loader
      summary: Create a new loader version (admin only)
      security:
        - bearerAuth: []

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  required: true
                  type: string
                  description: "Git tag that references the release"
                commit_hash:
                  required: true
                  type: string
                prerelease:
                  required: false
                  type: boolean
                gd:
                  $ref: "#/components/schemas/GDVersionObject"

      responses:
        "204":
          description: No Content (Version created)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/loader/versions/{version}:
    get:
      tags:
        - loader
      summary: Gets a loader version

      parameters:
        - name: version
          description: Geode Version
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ModVersionString"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoaderVersion"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: uuid
  schemas:
    ModID:
      type: string
      pattern: '[a-z0-9_\-]+\.[a-z0-9_\-]+'
      description: Mod ID
      examples:
        - geode.devtools
        - mat.circle-tool
    ModVersionString:
      type: string
      pattern: '^v?\d+\.\d+\.\d+(?:-.+)?$'
      examples:
        - "v1.0.0"
        - "v2.3.4"
        - "1.3.0"
        - "v1.0.0-beta"
    GDVersionString:
      type: string
      examples:
        - "2.200"
        - "2.204"
        - "2.205"
    GDVersionObject:
      type: object
      properties:
        win:
          anyOf:
            - type: "null"
            - $ref: "#/components/schemas/GDVersionString"
        mac:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/GDVersionString"
        ios:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/GDVersionString"
        android32:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/GDVersionString"
        android64:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/GDVersionString"
    ModUpdate:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ModID"
        version:
          $ref: "#/components/schemas/ModVersionString"
        download_link:
          type: string
          examples:
            - "https://api.geode-sdk.com/v1/mods/geode.nodeids/versions/1.0.0/download"
            - "https://api.geode-sdk.com/v1/mods/geode.devtools/versions/1.0.0/download"
        replacement:
          oneOf:
            - "null"
            - $ref: "#/components/schemas/ModReplacement"
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/ModDependency"
        incompatibilities:
          type: array
          items:
            $ref: "#/components/schemas/ModIncompatibility"
    ModReplacement:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ModID"
        version:
          $ref: "#/components/schemas/ModVersionString"
        download_link:
          type: string
          examples:
            - "https://api.geode-sdk.com/v1/mods/geode.nodeids/versions/1.0.0/download"
            - "https://api.geode-sdk.com/v1/mods/geode.devtools/versions/1.0.0/download"
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/ModDependency"
        incompatibilities:
          type: array
          items:
            $ref: "#/components/schemas/ModIncompatibility"
    ModDependency:
      type: object
      properties:
        mod_id:
          $ref: "#/components/schemas/ModID"
        version:
          type: string
          examples:
            - ">1.0.0"
            - "*"
            - "<=2.0.0"
        importance:
          $ref: "#/components/schemas/DependencyImportance"
    ModIncompatibility:
      type: object
      properties:
        mod_id:
          $ref: "#/components/schemas/ModID"
        version:
          type: string
          examples:
            - ">1.0.0"
            - "*"
            - "<=2.0.0"
        importance:
          $ref: "#/components/schemas/IncompatibilityImportance"
    Mod:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ModID"
        repository:
          oneOf:
            - type: "null"
            - type: string
          examples:
            - "https://github.com/geode-sdk/geode"
        developers:
          type: array
          items:
            $ref: "#/components/schemas/ModDeveloper"
        download_count:
          type: integer
        tags:
          type: array
          items:
            type: string
            examples:
              - "enhancement"
              - "gameplay"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/ModVersion"
        about:
          type: string
          examples:
            - "This mod adds bugfixes to the game"
        changelog:
          type: string
          examples:
            - "Fixed a bug with the mod"
      required:
        - id
    ModVersion:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          examples:
            - Devtools
        download_count:
          type: integer
        description:
          type: string
        early_load:
          type: boolean
        api:
          type: boolean
        geode:
          $ref: "#/components/schemas/ModVersionString"
        version:
          $ref: "#/components/schemas/ModVersionString"
        download_link:
          type: string
          examples:
            - "https://api.geode-sdk.com/v1/mods/geode.nodeids/versions/1.0.0/download"
            - "https://api.geode-sdk.com/v1/mods/geode.devtools/versions/1.0.0/download"
        hash:
          type: string
          description: This is generated serverside
          examples:
            - "3c8d6d3d48967758055a5569a24617c3e6fdc456fbf6a4adbf1222954e61b634"
        gd:
          description: The GD version the mod supports (can be specified per platform)
          oneOf:
            - $ref: "#/components/schemas/GDVersionString"
            - $ref: "#/components/schemas/GDVersionObject"
        tags:
          type: array
          items:
            type: string
        mod_id:
          $ref: "#/components/schemas/ModID"
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/ModDependency"
        incompatibilities:
          type: array
          items:
            $ref: "#/components/schemas/ModIncompatibility"
    ModVersionStatus:
      type: string
      enum:
        - accepted
        - rejected
        - pending
        - unlisted
    DependencyImportance:
      type: string
      enum:
        - suggested
        - recommended
        - required
    IncompatibilityImportance:
      type: string
      enum:
        - breaking
        - conflicting
        - superseded
    ModDeveloper:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
          description: The username of the developer. At the moment it's the same as the GitHub username of the developer.
          examples:
            - fleeym
        display_name:
          type: string
          examples:
            - Flame
        verified:
          type: boolean,
          description: Whether the developer can post mods without admin verification
          examples:
            - true
        admin:
          type: boolean,
          description: Whether the developer is an admin (which can verify mods and developers)
          examples:
            - true
        is_owner:
          type: boolean,
          description: Whether the developer is the owner of the mod
          examples:
            - true
...
    LoaderVersion:
      type: object
      properties:
        tag:
          type: string
        commit_hash:
          type: string
        version:
          $ref: "#/components/schemas/ModVersionString"
        prerelease:
          type: boolean
        gd:
          $ref: "#/components/schemas/GDVersionObject"
        created_at:
          type: string
          format: date-time
    Platform:
      type: string
      enum:
        - win
        - mac
        - android32
        - android64
        - ios
  parameters:
    ModID:
      name: id
      description: Mod ID
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ModID"
...
    GeodeVersionQuery:
      name: geode
      in: query
      description: Geode version being used
      required: false
      schema:
        $ref: "#/components/schemas/ModVersionString"
    Page:
      name: page
      in: query
      description: Page number
      required: false
      example: 1
      schema:
        type: integer
    PerPage:
      name: per_page
      in: query
      description: Number of elements to fetch per page
      required: false
      example: 10
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "You are not authorized to access this resource"
              payload:
                type: "null"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              payload:
                type: "null"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              payload:
                type: "null"
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              payload:
                type: "null"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              payload:
                type: "null"